+++ src/test/kotlin/db/tables/AudioMessagesUtil.kt
@@ -1,0 +1,5 @@
+package com.neelkamath.omniChat.db.tables
+
+import java.lang.ClassLoader.getSystemClassLoader
+
+fun buildMp3(fileName: String): Mp3 = getSystemClassLoader().getResource(fileName)!!.readBytes().let(::Mp3)