+++ src/test/kotlin/db/tables/PicsUtil.kt
@@ -1,0 +1,6 @@
+package com.neelkamath.omniChat.db.tables
+
+import java.lang.ClassLoader.getSystemClassLoader
+
+fun Pic.Companion.build(fileName: String): Pic =
+    getSystemClassLoader().getResource(fileName)!!.run { build(readBytes(), file.substringAfterLast(".")) }